+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 * 2016-06-06 Fixes off-click bug on scope destroy.
 * 2016-06-08 Fixes issue with nested expandables.
+* 2016-06-09 Adds ui-sref-active-for directive.